# static/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "AI Ticketing System API",
    "description": "Comprehensive AI-powered ticketing system with predictive analytics, knowledge base, and intelligent ticket management",
    "version": "1.0.0",
    "contact": {
      "name": "API Support"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Development server"
    }
  ],
...
  "paths": {
...
    "/analytics/forecast": {
      "get": {
        "tags": ["Analytics"],
        "summary": "Forecast ticket volume",
        "description": "Forecast total ticket volume for the next N days using exponential smoothing",
        "operationId": "forecast_volume",
        "parameters": [
          {
            "name": "days",
            "in": "query",
            "required": false,
            "description": "Number of historical days to use for forecasting",
            "schema": {
              "type": "integer",
              "default": 30
            }
          },
          {
            "name": "days_to_forecast",
            "in": "query",
            "required": false,
            "description": "Number of days to forecast into the future",
            "schema": {
              "type": "integer",
              "default": 7
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Forecast results with historical data and predictions",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "period": {
                      "type": "string"
                    },
                    "forecast_window": {
                      "type": "string"
                    },
                    "history": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "date": {
                            "type": "string",
                            "format": "date"
                          },
                          "count": {
                            "type": "integer"
                          }
                        }
                      }
                    },
                    "forecast": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "date": {
                            "type": "string",
                            "format": "date"
                          },
                          "count": {
                            "type": "integer"
                          }
                        }
                      }
                    },
                    "explanation": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/analytics/forecast-by-type": {
      "get": {
        "tags": ["Analytics"],
        "summary": "Forecast ticket volume by type",
        "description": "Forecast ticket volume broken down by type (Bug, Feature Request, Support, Task)",
        "operationId": "forecast_volume_by_type",
        "parameters": [
          {
            "name": "days",
            "in": "query",
            "required": false,
            "description": "Number of historical days to use for forecasting",
            "schema": {
              "type": "integer",
              "default": 30
            }
          },
          {
            "name": "days_to_forecast",
            "in": "query",
            "required": false,
            "description": "Number of days to forecast into the future",
            "schema": {
              "type": "integer",
              "default": 7
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Forecast results broken down by ticket type",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "period": {
                      "type": "string"
                    },
                    "forecast_window": {
                      "type": "string"
                    },
                    "history": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "date": {
                            "type": "string",
                            "format": "date"
                          },
                          "Bug": {
                            "type": "integer"
                          },
                          "Feature Request": {
                            "type": "integer"
                          },
                          "Support": {
                            "type": "integer"
                          },
                          "Task": {
                            "type": "integer"
                          },
                          "total": {
                            "type": "integer"
                          }
                        }
                      }
                    },
                    "forecast": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "date": {
                            "type": "string",
                            "format": "date"
                          },
                          "Bug": {
                            "type": "integer"
                          },
                          "Feature Request": {
                            "type": "integer"
                          },
                          "Support": {
                            "type": "integer"
                          },
                          "Task": {
                            "type": "integer"
                          },
                          "total": {
                            "type": "integer"
                          }
                        }
                      }
                    },
                    "breakdown": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Bearer token authentication"
      }
    }
  }
}